added resolved field in api and controller for topics, tested on nodebb server
