+++ src/controllers/write/topics.js
@@ -56,6 +56,12 @@ Topics.delete = async (req, res) => {
     helpers.formatApiResponse(200, res);
 };
 
+// added reosolved field
+Topics.resolved = async (req, res) => {
+    await api.topics.resolved(req, { tids: [req.params.tid] });
+    helpers.formatApiResponse(200, res);
+};
+
 Topics.restore = async (req, res) => {
     await api.topics.restore(req, { tids: [req.params.tid] });
     helpers.formatApiResponse(200, res);
